implement log out functionality
